feat(dialog): Add confirmation dialog for user choice
This commit adds a confirmation dialog to the user interface, prompting the user to confirm their choice. The `confirm()` function is used to display the dialog box with the message "Etes vous sûr(e) ?" (Are you sure?).

Please note that the commented-out lines with `alert()` and `prompt()` were not directly related to the functionality being implemented and have been removed.