chore(index): Add basic DOM manipulation code for testing
This commit adds the basic DOM manipulation code for testing purposes. The code changes can be found in the index.html file.

The added code includes:
- A JavaScript file "script.js" linked with the "defer" attribute.
- HTML structure with headings, sections, lists, and paragraphs.

This commit is a chore since it doesn't introduce any new feature or fix any bug. It is solely for the purpose of adding code for testing DOM manipulation.

Note: The commit message assumes that you are following the Angular commit convention. If you're using a different convention, you may modify the commit message accordingly.

References:
- Angular Commit Message Guidelines: https://github.com/angular/angular/blob/master/CONTRIBUTING.md#commit