chore(index): Remove defer attribute from script tag in index.html
This commit removes the defer attribute from the script tag in the index.html file. The change does not affect the behavior of the program.

The defer attribute was removed to ensure the script is executed immediately when encountered by the browser, instead of waiting for the HTML parsing and rendering to complete.

This is a chore commit as it does not introduce any functional changes but improves the code structure or maintainability.

Note: The commit message assumes that you are following the Angular commit convention. If you're using a different convention, you may modify the commit message accordingly.

References:
- Angular Commit Message Guidelines: https://github.com/angular/angular/blob/master/CONTRIBUTING.md#commit