refactor: rename folder and HTML file
Renamed the folder containing the JavaScript program and the associated HTML file for better organization and clarity.

The folder "CH2_1" has been changed to "Exercice 2 - Surfaces de cercles" and the file "CH2_1.html" is now "index.html."

No changes have been made to the content or functionality of the program itself.